Moving low-level ZK storage classes to the new `persistence-zookeeper` module
Additionally:
- Removed the `TemplateRepository`, `PathTrie` and other concrete repository classes. Let's get aligned on how the repository interface looks like first and add a concrete implementation in a separate PR
- removed unnecessary dependency libs like `wix-accord` and moved some testing classes to `test-utils` module
