+++ persistence-zookeeper/build.gradle
@@ -1,0 +1,7 @@
+dependencies {
+    compile project(':core-models')
+    implementation project(':metrics')
+    implementation project(':test-utils')
+
+    compile "org.scala-lang.modules:scala-async_$scalaVersion:0.9.7"
+}
