+++ persistence/build.gradle
@@ -1,13 +1,3 @@
 dependencies {
     compile project(':core-models')
-    implementation project(':metrics')
-    implementation project(':test-utils')
-
-    compile "org.scala-lang.modules:scala-async_$scalaVersion:0.9.7"
-    compile "com.wix:accord-core_$scalaVersion:0.7.1"
-
-    testCompile "com.typesafe.akka:akka-testkit_$scalaVersion:2.5.14"
-    testCompile "org.scalatest:scalatest_$scalaVersion:3.0.4"
-    testCompile "org.mockito:mockito-all:1.10.19"
-    testCompile "com.typesafe.akka:akka-testkit_$scalaVersion:2.5.14"
 }
